Updated addon to 1.1 for Tamriel_Data v10
Changed naming to "Tamriel Data" from Tamriel Rebuilt. Empty files in the TamrielRebuilt folder were included to ensure that only the new version of the addon is run.
Added fixPlayerRaceAnimations feature to allow players to have the intended animations (with tails) when playing as Ohmes-raht and Suthay Khajiit via 3rd party mods. This feature can easily be expanded to account for future races which will use their own skeletons/animations, such as the Tsaesci.
Registered the messagebox telling people to register their BSAs as part of the "loaded" event. Previously, the addon would assume that the game should (be able to) show this messagebox before any files were loaded, resulting in a crash.
Updated MWSE buildDate.
Added my name to the credits.