Added function improving behavior of vampire heads
Namira's shroud will now show the player's non-vampiric head rather than removing the head entirely. If the player repeatedly equips and unequips gear in the inventory screen, then sometimes this will fail for reasons that are not clear to me, but that is easy to fix by just equipping or unequipping a piece of equipment. Vampires who are supposed to have unique heads (e.g. Varos) can also be accounted for in this function so that they aren't forced to use generic vampire heads.

Some improvements were made to the way that event registrations were also made.