Modernized MCM/Config and added transparency check
The MCM template is now made in a function that is registered to the modConfigReady event.

When using passwall on an activator, the targeted mesh will be checked for transparency/alpha blending, which should prevent it from being used to bypass forcefields, particles, and other immaterial objects that it either wouldn't make sense to or cause problems with quests.

The rock chisel clannfear summon has been renamed per Cicero's recent changes.

The "tr" spell tables have had "tr" replaced with "td".

tes3.addMagicEffect no longer iterates through the td_miscs table, since the spells in it will differ from each other much more than the summons do.

config.lua now sets the options to their default value if they do not already exist in the user's config file, so they don't have to manually enable stuff in MCM when they update to a version of Tamriel_Data with new features.