9/8/2024 Merge
Might as well start dating merges that are as large as this one.

Wrote code to change TD enchantments and spells other than the main ones that just have a single replaceable effect. Various summon scrolls now exist for most of the new summons and Kyne's intervention; the Ayleid sorcerer king will also summon suitable creature instead of just vanilla daedra now.
Added feature to replace the effects of Magicka Expanded spells. The cost of the spells is not changed, because MWSE's function for calculating the cost does not work with them for some reason.
Adjusted the cost of the newer summon spells with a new formula that closely matches the values that mort calculated.
Used tes3.effect.* for the IDs of new spells rather than just the number for the sake of clarity and consistency.
Various minor changes to the text in the code and comments.