Made tanthas attack the player when looting a nest
Also disabled changing the Morrowind reputation when outside of Morrowind, since it was not working consistently and greatly increasing Morrowind reputation as a result. The Tantha nests were added to the GH blacklist for now, since there is a lot of loot that one might not expect to get when looting one without there being any visual changes to the nests themselves.