Fixed distracted actors not moving in some cases
The three closest pathgrid nodes were not considered valid destinations even when they were the only possible ones. Also removed a poorly-suited file from distractedVoiceLines , prevented vampires from saying the non-vampiric lines and a possible error from data.tamrielData not existing on a distracted target that was not able to find a destination.